app[ARCHI](navigation) added navigation folder
