--- packages/app/src/Resources/Navigation.ts
@@ -1,14 +1,0 @@
-export enum Tabs {
-  home = "HOME",
-  cart = "CART",
-}
-
-export enum Screens {
-  home = "Home",
-  cart = "Cart",
-  signIn = "SignIn",
-  signUp = "SignUp",
-  loggedHome = "LoggedHome",
-  pay = "Pay",
-  account = "Account",
-}
